Extractor code cleanup v1
