--- detector_parameters.yaml
@@ -1,213 +1,0 @@
-/**:
-  detector:
-    kstrongest:
-      parameter_count: 1
-      parameters:
-        kstrong:
-          min: 1
-          max: 6
-          step: 1
-
-    modified_cacfar_power:
-      parameter_count: 1
-      parameters:
-        threshold:
-          # min: 20.0
-          # max: 120.0
-          # step: 5.0
-          min: 90.0
-          max: 150.0
-          step: 10.0
-      #16.5,24
-    cago_cfar:
-      parameter_count: 1
-      parameters:
-        threshold:
-          min: 10.0
-          max: 130.0
-          step: 10.0
-          # min: 25.0
-          # max: 25.0
-          # step: 1.0
-        # static_threshold:
-        #   min: 10.0
-        #   max: 15.0
-        #   step: 1.0
-    caso_cfar:
-      parameter_count: 1
-      parameters:
-        threshold:
-          min: 3100.0
-          max: 4700.0
-          step: 200.0
-          # min: 400.0
-          # max: 400.0
-          # step: 1.0
-        # static_threshold:
-        #   min: 10.0
-        #   max: 15.0
-        #   step: 1.0
-    cfear_kstrong:
-      parameter_count: 2
-      parameters:
-        kstrong:
-          min: 20
-          max: 20
-          step: 4
-        r: 
-          min: 0.5
-          max: 0.5
-          step: 0.5
-        # kstrong:
-        #   min: 12
-        #   max: 24
-        #   step: 4
-        # r: 
-        #   min: 0.5
-        #   max: 1.0
-        #   step: 0.5
-        
-
-    bfar_pure:
-      parameter_count: 2
-      parameters:
-        threshold:
-          min: 5.0
-          max: 20.0
-          step: 5.0
-        threshold3:
-          min: 0.2
-          max: 0.35
-          step: 0.05
-        # threshold:
-        #   min: 10.0
-        #   max: 50.0
-        #   step: 10.0
-        # threshold3:
-        #   min: 0.05
-        #   max: 0.3
-        #   step: 0.05
-    oscfar:
-      parameter_count: 1
-      parameters:
-          threshold: 
-            min: 120.0
-            max: 120.0
-            step: 20.0
-            # min: 20.0
-            # max: 130.0
-            # step: 5.0
-    tm_cfar:
-      parameter_count: 2
-      parameters:
-        threshold:
-          min: 100.0
-          max: 100.0 
-          step: 20.0
-          # min: 40.0
-          # max: 100.0 #100.0
-          # step: 20.0
-        N1:
-          min: 30
-          max: 30
-          step: 4
-          # min: 40
-          # max: 48
-          # step: 4
-    is_cfar:
-      parameter_count: 3
-      parameters:
-        alpha_I: #[0.01,10.0]
-          min: 0.075
-          max: 0.075
-          step: 5.0
-        # N_TI: # N_TI = N - N_i - 1 = 49  N_i, N_i = [49,41]
-        #   min: 43
-        #   max: 43
-        #   step: 4
-        beta_I:
-          # min: 15.0
-          # max: 20.0
-          # step: 5.0
-          min: 15.0
-          max: 15.0
-          step: 5.0
-        # beta_I:
-        #   min: 0.0
-        #   max: 0.25
-        #   step: 0.05
-
-    # NOTE ABOUT TUNING: one thing about the scaling threshold C_N is that a different weighting onCAGO vs CASO will require different
-    # Thresholds to optimize. One idea could be select the best threshold from CASO, CAGO etc and lock it in as opposed to using the same threshold fro each one
-    vi_cfar:
-      # parameter_count: 3
-      parameter_count: 2
-      parameters:
-        K_VI: #[2.020.0][entierly CASO  Mostly CA/CAGO]
-          # min: 1.0
-          # max: 10.0
-          # step: 4.0
-          min: 5.0
-          max: 5.0
-          step: 3.0
-        # K_MR: #1.5 [1.03] makes sense to cap around 3, cant say the means are the same if one is 3 times bigger
-        #   min: 1.5
-        #   max: 2.5
-        #   step: 0.5
-        # C_N: # Huge range [201000]
-        #   min: 80.0
-        #   max: 80.0
-        #   step: 30.0
-        C_N: # Huge range [201000]
-          min: 400.0
-          max: 400.0
-          step: 200.0
-    msca_cfar:
-      parameter_count: 2
-      parameters:
-        threshold:
-          # min: 30.0
-          # max: 100.0
-          # step: 10.0
-          min: 100.0
-          max: 100.0
-          step: 20.0
-        M:
-          min: 8
-          max: 8
-          step: 2
-
-        
-    landmark_extraction:
-      parameter_count: 3
-      parameters:
-        w_median:
-          min: 50
-          max: 250
-          step: 50
-        w_binom:
-          min: 3
-          max: 10
-          step: 3
-        z_q:
-          min: 1.5
-          max: 3.5
-          step: 1
-    cen2018:
-      parameter_count: 1
-      parameters:
-        sigma:
-          min: 10
-          max: 10
-          step: 2
-        zq:
-          min: 2.75
-          max: 2.75
-          step: 0.25
-    cen2019:
-      parameter_count: 1
-      parameters:
-        l_max:
-          min: 100
-          max: 200
-          step: 100
